Split Git loader into two separate ones
